scripts/requirements.txt
# 域名知识体系项目依赖包

# DNS查询和解析
dnspython>=2.3.0

# Whois查询
python-whois>=0.7.3

# HTTP请求
requests>=2.28.0

# 数据处理
pandas>=1.5.0
numpy>=1.24.0

# 命令行工具
click>=8.1.0
argparse

# 异步处理
asyncio
aiohttp>=3.8.0

# 网络工具
socket
ipaddress

# 时间处理
datetime
time

# 文件处理
csv
json
yaml>=6.0

# 正则表达式
re

# 系统工具
os
sys
subprocess

# 并发处理
concurrent.futures
threading

# 加密和安全
cryptography>=3.4.8
ssl

# 监控和日志
logging
psutil>=5.9.0

# 可视化（可选）
matplotlib>=3.6.0
seaborn>=0.11.0

# 数据库（可选）
sqlite3
sqlalchemy>=1.4.0

# 配置文件处理
configparser
toml>=0.10.0

# 测试工具（开发用）
pytest>=7.0.0
pytest-asyncio>=0.21.0